PHPC-910: Split BSON encoding/decoding and prefix symbols
