+++ src/MongoDB/Cursor.c
@@ -125,7 +125,7 @@ static void php_phongo_cursor_iterator_move_forward(zend_object_iterator *iter T
 	cursor->current++;
 
 	if (mongoc_cursor_next(cursor->cursor, &doc)) {
-		phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &cursor->visitor_data);
+		php_phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &cursor->visitor_data);
 	} else {
 		bson_error_t error;
 
@@ -153,7 +153,7 @@ static void php_phongo_cursor_iterator_rewind(zend_object_iterator *iter TSRMLS_
 	doc = mongoc_cursor_current(cursor->cursor);
 
 	if (doc) {
-		phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &cursor->visitor_data);
+		php_phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &cursor->visitor_data);
 	}
 } /* }}} */
 
@@ -227,7 +227,7 @@ PHP_METHOD(Cursor, setTypeMap)
 		php_phongo_cursor_free_current(intern);
 	}
 
-	phongo_bson_typemap_to_state(typemap, &state.map TSRMLS_CC);
+	php_phongo_bson_typemap_to_state(typemap, &state.map TSRMLS_CC);
 
 	intern->visitor_data = state;
 
@@ -236,7 +236,7 @@ PHP_METHOD(Cursor, setTypeMap)
 	if (mongoc_cursor_current(intern->cursor)) {
 		const bson_t *doc = mongoc_cursor_current(intern->cursor);
 
-		phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &intern->visitor_data);
+		php_phongo_bson_to_zval_ex(bson_get_data(doc), doc->len, &intern->visitor_data);
 	}
 }
 /* }}} */
