+++ src/MongoDB/Server.c
@@ -190,7 +190,7 @@ PHP_METHOD(Server, getTags)
 			state.map.document_type = PHONGO_TYPEMAP_NATIVE_ARRAY;
 
 			bson_iter_document(&iter, &len, &bytes);
-			phongo_bson_to_zval_ex(bytes, len, &state);
+			php_phongo_bson_to_zval_ex(bytes, len, &state);
 			mongoc_server_description_destroy(sd);
 
 #if PHP_VERSION_ID >= 70000
@@ -232,7 +232,7 @@ PHP_METHOD(Server, getInfo)
 		state.map.root_type = PHONGO_TYPEMAP_NATIVE_ARRAY;
 		state.map.document_type = PHONGO_TYPEMAP_NATIVE_ARRAY;
 
-		phongo_bson_to_zval_ex(bson_get_data(is_master), is_master->len, &state);
+		php_phongo_bson_to_zval_ex(bson_get_data(is_master), is_master->len, &state);
 		mongoc_server_description_destroy(sd);
 
 #if PHP_VERSION_ID >= 70000
