+++ src/bson.c
@@ -18,176 +18,39 @@
 #	include "config.h"
 #endif
 
-/* External libs */
 #include <bson.h>
 
-/* PHP Core stuff */
 #include <php.h>
-#include <ext/spl/spl_array.h>
 #include <Zend/zend_hash.h>
 #include <Zend/zend_interfaces.h>
 
-/* PHP array helpers */
-#include "php_array_api.h"
-
-/* Our Compatability header */
-#include "phongo_compat.h"
-
-/* Our stuffz */
 #include "php_phongo.h"
 #include "php_bson.h"
-
-#if SIZEOF_PHONGO_LONG == 8
-# define BSON_APPEND_INT(b, key, keylen, val) \
-	if (val > INT32_MAX || val < INT32_MIN) { \
-		bson_append_int64(b, key, keylen, val); \
-	} else { \
-		bson_append_int32(b, key, keylen, val); \
-	}
-#elif SIZEOF_PHONGO_LONG == 4
-# define BSON_APPEND_INT(b, key, keylen, val) \
-	bson_append_int32(b, key, keylen, val)
-#else
-# error Unsupported architecture (integers are neither 32-bit nor 64-bit)
-#endif
+#include "phongo_compat.h"
+#include "php_array_api.h"
 
 #undef MONGOC_LOG_DOMAIN
 #define MONGOC_LOG_DOMAIN "PHONGO-BSON"
-
-#define PHONGO_ODM_FIELD_NAME "__pclass"
 
 #define PHONGO_IS_CLASS_INSTANTIATABLE(ce) \
 	(!(ce->ce_flags & (ZEND_ACC_INTERFACE|ZEND_ACC_IMPLICIT_ABSTRACT_CLASS|ZEND_ACC_EXPLICIT_ABSTRACT_CLASS)))
 
 /* Forward declarations */
-bool php_phongo_bson_visit_document(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data);
-bool php_phongo_bson_visit_array(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data);
+static bool php_phongo_bson_visit_document(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data);
+static bool php_phongo_bson_visit_array(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data);
 
-/* {{{ Santa's Little Helper: Object getters */
-void php_phongo_objectid_get_id(zval *object, bson_oid_t *oid TSRMLS_DC)
-{
-	php_phongo_objectid_t     *intern;
-
-	intern = Z_OBJECTID_OBJ_P(object);
-
-	bson_oid_init_from_string(oid, intern->oid);
-}
-int64_t php_phongo_utcdatetime_get_milliseconds(zval *object TSRMLS_DC)
-{
-	php_phongo_utcdatetime_t     *intern;
-
-	intern = Z_UTCDATETIME_OBJ_P(object);
-
-	return intern->milliseconds;
-}
-uint32_t php_phongo_timestamp_get_increment(zval *object TSRMLS_DC)
-{
-	php_phongo_timestamp_t     *intern;
-
-	intern = Z_TIMESTAMP_OBJ_P(object);
-
-	return intern->increment;
-}
-uint32_t php_phongo_timestamp_get_timestamp(zval *object TSRMLS_DC)
-{
-	php_phongo_timestamp_t     *intern;
-
-	intern = Z_TIMESTAMP_OBJ_P(object);
-
-	return intern->timestamp;
-}
-bool php_phongo_javascript_has_scope(zval *object TSRMLS_DC)
-{
-	php_phongo_javascript_t *intern;
-
-	intern = Z_JAVASCRIPT_OBJ_P(object);
-
-	return !!intern->scope;
-}
-char *php_phongo_javascript_get_code(zval *object TSRMLS_DC)
-{
-	php_phongo_javascript_t *intern;
-
-	intern = Z_JAVASCRIPT_OBJ_P(object);
-
-	return intern->code;
-}
-bson_t *php_phongo_javascript_get_scope(zval *object TSRMLS_DC)
-{
-	php_phongo_javascript_t *intern;
-
-	intern = Z_JAVASCRIPT_OBJ_P(object);
-
-	return intern->scope;
-}
-int php_phongo_binary_get_data(zval *object, char **data TSRMLS_DC)
-{
-	php_phongo_binary_t *intern;
-
-	intern = Z_BINARY_OBJ_P(object);
-
-	*data = intern->data;
-	return intern->data_len;
-}
-uint8_t php_phongo_binary_get_type(zval *object TSRMLS_DC)
-{
-	php_phongo_binary_t *intern;
-
-	intern = Z_BINARY_OBJ_P(object);
-
-	return intern->type;
-}
-
-static bson_decimal128_t *php_phongo_decimal128_get_decimal128(zval *object TSRMLS_DC)
-{
-	php_phongo_decimal128_t *intern;
-
-	intern = Z_DECIMAL128_OBJ_P(object);
-
-	return &intern->decimal;
-}
-
-char *php_phongo_regex_get_pattern(zval *object TSRMLS_DC)
-{
-	php_phongo_regex_t *intern;
-
-	intern = Z_REGEX_OBJ_P(object);
-
-	return intern->pattern;
-}
-char *php_phongo_regex_get_flags(zval *object TSRMLS_DC)
-{
-	php_phongo_regex_t *intern;
-
-	intern = Z_REGEX_OBJ_P(object);
-
-	return intern->flags;
-}
-/* }}} */
-#if 0
-bool php_phongo_bson_visit_before(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
-{
-	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_after(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
-{
-	return false;
-}
-/* }}} */
-#endif
-void php_phongo_bson_visit_corrupt(const bson_iter_t *iter ARG_UNUSED, void *data ARG_UNUSED) /* {{{ */
+static void php_phongo_bson_visit_corrupt(const bson_iter_t *iter ARG_UNUSED, void *data ARG_UNUSED) /* {{{ */
 {
 	mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "Corrupt BSON data detected!");
-}
+} /* }}} */
 
-void php_phongo_bson_visit_unsupported_type(const bson_iter_t *iter ARG_UNUSED, const char *key, uint32_t v_type_code, void *data ARG_UNUSED) /* {{{ */
+static void php_phongo_bson_visit_unsupported_type(const bson_iter_t *iter ARG_UNUSED, const char *key, uint32_t v_type_code, void *data ARG_UNUSED) /* {{{ */
 {
 	TSRMLS_FETCH();
 	phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "Detected unknown BSON type 0x%02hhx for fieldname \"%s\". Are you using the latest driver?", v_type_code, key);
-}
+} /* }}} */
 
-bool php_phongo_bson_visit_double(const bson_iter_t *iter ARG_UNUSED, const char *key, double v_double, void *data) /* {{{ */
+static bool php_phongo_bson_visit_double(const bson_iter_t *iter ARG_UNUSED, const char *key, double v_double, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -202,9 +65,9 @@ bool php_phongo_bson_visit_double(const bson_iter_t *iter ARG_UNUSED, const char
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_utf8(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_utf8_len, const char *v_utf8, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_utf8(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_utf8_len, const char *v_utf8, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -219,9 +82,9 @@ bool php_phongo_bson_visit_utf8(const bson_iter_t *iter ARG_UNUSED, const char *
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_binary(const bson_iter_t *iter ARG_UNUSED, const char *key, bson_subtype_t v_subtype, size_t v_binary_len, const uint8_t *v_binary, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_binary(const bson_iter_t *iter ARG_UNUSED, const char *key, bson_subtype_t v_subtype, size_t v_binary_len, const uint8_t *v_binary, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -272,19 +135,9 @@ bool php_phongo_bson_visit_binary(const bson_iter_t *iter ARG_UNUSED, const char
 	}
 
 	return false;
-}
-/* }}} */
-#if 0
-bool php_phongo_bson_visit_undefined(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
-{
-	printf("Not Implemented\n");
+} /* }}} */
 
-	return true;
-}
-/* }}} */
-#endif
-
-bool php_phongo_bson_visit_oid(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_oid_t *v_oid, void *data) /* {{{ */
+static bool php_phongo_bson_visit_oid(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_oid_t *v_oid, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -317,9 +170,9 @@ bool php_phongo_bson_visit_oid(const bson_iter_t *iter ARG_UNUSED, const char *k
 #endif
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_bool(const bson_iter_t *iter ARG_UNUSED, const char *key, bool v_bool, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_bool(const bson_iter_t *iter ARG_UNUSED, const char *key, bool v_bool, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -334,9 +187,9 @@ bool php_phongo_bson_visit_bool(const bson_iter_t *iter ARG_UNUSED, const char *
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_date_time(const bson_iter_t *iter ARG_UNUSED, const char *key, int64_t msec_since_epoch, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_date_time(const bson_iter_t *iter ARG_UNUSED, const char *key, int64_t msec_since_epoch, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -369,10 +222,9 @@ bool php_phongo_bson_visit_date_time(const bson_iter_t *iter ARG_UNUSED, const c
 #endif
 
 	return false;
-}
-/* }}} */
+} /* }}} */
 
-bool php_phongo_bson_visit_decimal128(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_decimal128_t *decimal, void *data) /* {{{ */
+static bool php_phongo_bson_visit_decimal128(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_decimal128_t *decimal, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -402,10 +254,9 @@ bool php_phongo_bson_visit_decimal128(const bson_iter_t *iter ARG_UNUSED, const 
 #endif
 
 	return false;
-}
-/* }}} */
+} /* }}} */
 
-bool php_phongo_bson_visit_null(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
+static bool php_phongo_bson_visit_null(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -420,9 +271,9 @@ bool php_phongo_bson_visit_null(const bson_iter_t *iter ARG_UNUSED, const char *
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_regex(const bson_iter_t *iter ARG_UNUSED, const char *key, const char *v_regex, const char *v_options, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_regex(const bson_iter_t *iter ARG_UNUSED, const char *key, const char *v_regex, const char *v_options, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -455,18 +306,9 @@ bool php_phongo_bson_visit_regex(const bson_iter_t *iter ARG_UNUSED, const char 
 #endif
 
 	return false;
-}
-/* }}} */
-#if 0
-bool php_phongo_bson_visit_dbpointer(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_collection_len, const char *v_collection, const bson_oid_t *v_oid, void *data) /* {{{ */
-{
-	printf("Not Implemented\n");
+} /* }}} */
 
-	return true;
-}
-/* }}} */
-#endif
-bool php_phongo_bson_visit_code(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_code_len, const char *v_code, void *data) /* {{{ */
+static bool php_phongo_bson_visit_code(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_code_len, const char *v_code, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -499,18 +341,9 @@ bool php_phongo_bson_visit_code(const bson_iter_t *iter ARG_UNUSED, const char *
 #endif
 
 	return false;
-}
-/* }}} */
-#if 0
-bool php_phongo_bson_visit_symbol(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_symbol_len, const char *v_symbol, void *data) /* {{{ */
-{
-	printf("Not Implemented\n");
+} /* }}} */
 
-	return true;
-}
-/* }}} */
-#endif
-bool php_phongo_bson_visit_codewscope(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_code_len, const char *v_code, const bson_t *v_scope, void *data) /* {{{ */
+static bool php_phongo_bson_visit_codewscope(const bson_iter_t *iter ARG_UNUSED, const char *key, size_t v_code_len, const char *v_code, const bson_t *v_scope, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -543,9 +376,9 @@ bool php_phongo_bson_visit_codewscope(const bson_iter_t *iter ARG_UNUSED, const 
 #endif
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_int32(const bson_iter_t *iter ARG_UNUSED, const char *key, int32_t v_int32, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_int32(const bson_iter_t *iter ARG_UNUSED, const char *key, int32_t v_int32, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -560,9 +393,9 @@ bool php_phongo_bson_visit_int32(const bson_iter_t *iter ARG_UNUSED, const char 
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_timestamp(const bson_iter_t *iter ARG_UNUSED, const char *key, uint32_t v_timestamp, uint32_t v_increment, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_timestamp(const bson_iter_t *iter ARG_UNUSED, const char *key, uint32_t v_timestamp, uint32_t v_increment, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -595,9 +428,9 @@ bool php_phongo_bson_visit_timestamp(const bson_iter_t *iter ARG_UNUSED, const c
 #endif
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_int64(const bson_iter_t *iter ARG_UNUSED, const char *key, int64_t v_int64, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_int64(const bson_iter_t *iter ARG_UNUSED, const char *key, int64_t v_int64, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -615,9 +448,9 @@ bool php_phongo_bson_visit_int64(const bson_iter_t *iter ARG_UNUSED, const char 
 	}
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_maxkey(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_maxkey(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -650,9 +483,9 @@ bool php_phongo_bson_visit_maxkey(const bson_iter_t *iter ARG_UNUSED, const char
 #endif
 
 	return false;
-}
-/* }}} */
-bool php_phongo_bson_visit_minkey(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
+} /* }}} */
+
+static bool php_phongo_bson_visit_minkey(const bson_iter_t *iter ARG_UNUSED, const char *key, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -685,9 +518,7 @@ bool php_phongo_bson_visit_minkey(const bson_iter_t *iter ARG_UNUSED, const char
 #endif
 
 	return false;
-}
-/* }}} */
-
+} /* }}} */
 
 static const bson_visitor_t php_bson_visitors = {
    NULL /* php_phongo_bson_visit_before*/,
@@ -718,8 +549,7 @@ static const bson_visitor_t php_bson_visitors = {
    { NULL }
 };
 
-
-bool php_phongo_bson_visit_document(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data)
+static bool php_phongo_bson_visit_document(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_document, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -825,10 +655,9 @@ bool php_phongo_bson_visit_document(const bson_iter_t *iter ARG_UNUSED, const ch
 	}
 
 	return false;
-}
+} /* }}} */
 
-
-bool php_phongo_bson_visit_array(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_array, void *data)
+static bool php_phongo_bson_visit_array(const bson_iter_t *iter ARG_UNUSED, const char *key, const bson_t *v_array, void *data) /* {{{ */
 {
 #if PHP_VERSION_ID >= 70000
 	zval *retval = &((php_phongo_bson_state *)data)->zchild;
@@ -937,566 +766,19 @@ bool php_phongo_bson_visit_array(const bson_iter_t *iter ARG_UNUSED, const char 
 	}
 
 	return false;
-}
+} /* }}} */
 
-
-int php_phongo_is_array_or_document(zval *val TSRMLS_DC) /* {{{ */
-{
-	HashTable *ht_data = HASH_OF(val);
-	int        count;
-
-	if (Z_TYPE_P(val) != IS_ARRAY) {
-		return IS_OBJECT;
-	}
-
-	count = ht_data ? zend_hash_num_elements(ht_data) : 0;
-	if (count > 0) {
+/* Converts a BSON document to a PHP value using the default typemap. */
 #if PHP_VERSION_ID >= 70000
-		zend_string *key;
-		zend_ulong index, idx;
-
-		idx = 0;
-		ZEND_HASH_FOREACH_KEY(ht_data, index, key) {
-			if (key) {
-				return IS_OBJECT;
-			} else {
-				if (index != idx) {
-					return IS_OBJECT;
-				}
-			}
-			idx++;
-		} ZEND_HASH_FOREACH_END();
+int php_phongo_bson_to_zval(const unsigned char *data, int data_len, zval *zv) /* {{{ */
 #else
-		char          *key;
-		unsigned int   key_len;
-		unsigned long  index = 0;
-		unsigned long  idx = 0;
-		int            hash_type = 0;
-		HashPosition   pos;
-
-		zend_hash_internal_pointer_reset_ex(ht_data, &pos);
-		for (;; zend_hash_move_forward_ex(ht_data, &pos)) {
-			hash_type = zend_hash_get_current_key_ex(ht_data, &key, &key_len, &index, 0, &pos);
-			if (hash_type == HASH_KEY_NON_EXISTENT) {
-				break;
-			}
-
-			if (hash_type == HASH_KEY_IS_STRING) {
-				return IS_OBJECT;
-			} else {
-				if (index != idx) {
-					return IS_OBJECT;
-				}
-			}
-			idx++;
-		}
-#endif
-	} else {
-		return Z_TYPE_P(val);
-	}
-
-	return IS_ARRAY;
-}
-/* }}} */
-void object_to_bson(zval *object, php_phongo_bson_flags_t flags, const char *key, long key_len, bson_t *bson TSRMLS_DC)
-{
-	bson_t child;
-
-	if (Z_TYPE_P(object) == IS_OBJECT && instanceof_function(Z_OBJCE_P(object), php_phongo_type_ce TSRMLS_CC)) {
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_serializable_ce TSRMLS_CC)) {
-#if PHP_VERSION_ID >= 70000
-			zval obj_data;
-#else
-			zval *obj_data = NULL;
-#endif
-			bson_t child;
-			HashTable *tmp_ht;
-#if PHP_VERSION_ID >= 70000
-			zend_call_method_with_0_params(object, NULL, NULL, BSON_SERIALIZE_FUNC_NAME, &obj_data);
-#else
-			zend_call_method_with_0_params(&object, NULL, NULL, BSON_SERIALIZE_FUNC_NAME, &obj_data);
-#endif
-
-			if (Z_ISUNDEF(obj_data)) {
-				/* zend_call_method() failed */
-				return;
-			}
-
-#if PHP_VERSION_ID >= 70000
-			if (Z_TYPE(obj_data) != IS_ARRAY && !(Z_TYPE(obj_data) == IS_OBJECT && instanceof_function(Z_OBJCE(obj_data), zend_standard_class_def TSRMLS_CC))) {
-				phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC,
-					"Expected %s::%s() to return an array or stdClass, %s given",
-					Z_OBJCE_P(object)->name->val,
-					BSON_SERIALIZE_FUNC_NAME,
-					(Z_TYPE(obj_data) == IS_OBJECT
-						 ? Z_OBJCE(obj_data)->name->val
-						 : zend_get_type_by_const(Z_TYPE(obj_data))
-					)
-				);
-				zval_ptr_dtor(&obj_data);
-#else
-			if (Z_TYPE_P(obj_data) != IS_ARRAY && !(Z_TYPE_P(obj_data) == IS_OBJECT && instanceof_function(Z_OBJCE_P(obj_data), zend_standard_class_def TSRMLS_CC))) {
-				phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC,
-					"Expected %s::%s() to return an array or stdClass, %s given",
-					Z_OBJCE_P(object)->name,
-					BSON_SERIALIZE_FUNC_NAME,
-					(Z_TYPE_P(obj_data) == IS_OBJECT
-						 ? Z_OBJCE_P(obj_data)->name
-						 : zend_get_type_by_const(Z_TYPE_P(obj_data))
-					)
-				);
-				zval_ptr_dtor(&obj_data);
-#endif
-
-				return;
-			}
-
-#if PHP_VERSION_ID >= 70000
-			tmp_ht = HASH_OF(&obj_data);
-#else
-			tmp_ht = HASH_OF(obj_data);
-#endif
-
-			if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-				ZEND_HASH_INC_APPLY_COUNT(tmp_ht);
-			}
-
-			/* Persistable objects must always be serialized as BSON documents;
-			 * otherwise, infer based on bsonSerialize()'s return value. */
-#if PHP_VERSION_ID >= 70000
-			if (instanceof_function(Z_OBJCE_P(object), php_phongo_persistable_ce TSRMLS_CC) || php_phongo_is_array_or_document(&obj_data TSRMLS_CC) == IS_OBJECT) {
-#else
-			if (instanceof_function(Z_OBJCE_P(object), php_phongo_persistable_ce TSRMLS_CC) || php_phongo_is_array_or_document(obj_data TSRMLS_CC) == IS_OBJECT) {
-#endif
-				bson_append_document_begin(bson, key, key_len, &child);
-				if (instanceof_function(Z_OBJCE_P(object), php_phongo_persistable_ce TSRMLS_CC)) {
-#if PHP_VERSION_ID >= 70000
-					bson_append_binary(&child, PHONGO_ODM_FIELD_NAME, -1, 0x80, (const uint8_t *)Z_OBJCE_P(object)->name->val, Z_OBJCE_P(object)->name->len);
-#else
-					bson_append_binary(&child, PHONGO_ODM_FIELD_NAME, -1, 0x80, (const uint8_t *)Z_OBJCE_P(object)->name, strlen(Z_OBJCE_P(object)->name));
-#endif
-				}
-#if PHP_VERSION_ID >= 70000
-				phongo_zval_to_bson(&obj_data, flags, &child, NULL TSRMLS_CC);
-#else
-				phongo_zval_to_bson(obj_data, flags, &child, NULL TSRMLS_CC);
-#endif
-				bson_append_document_end(bson, &child);
-			} else {
-				bson_append_array_begin(bson, key, key_len, &child);
-#if PHP_VERSION_ID >= 70000
-				phongo_zval_to_bson(&obj_data, flags, &child, NULL TSRMLS_CC);
-#else
-				phongo_zval_to_bson(obj_data, flags, &child, NULL TSRMLS_CC);
-#endif
-				bson_append_array_end(bson, &child);
-			}
-
-			if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-				ZEND_HASH_DEC_APPLY_COUNT(tmp_ht);
-			}
-			zval_ptr_dtor(&obj_data);
-			return;
-		}
-
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_objectid_ce TSRMLS_CC)) {
-			bson_oid_t oid;
-
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding ObjectId");
-			php_phongo_objectid_get_id(object, &oid TSRMLS_CC);
-			bson_append_oid(bson, key, key_len, &oid);
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_utcdatetime_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding UTCDateTime");
-			bson_append_date_time(bson, key, key_len, php_phongo_utcdatetime_get_milliseconds(object TSRMLS_CC));
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_binary_ce TSRMLS_CC)) {
-			const unsigned char *data;
-			size_t data_len;
-
-			data_len = php_phongo_binary_get_data(object, (char **)&data TSRMLS_CC);
-
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Binary");
-			bson_append_binary(bson, key, key_len, php_phongo_binary_get_type(object TSRMLS_CC), data, data_len);
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_decimal128_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Decimal128");
-			bson_append_decimal128(bson, key, key_len, php_phongo_decimal128_get_decimal128(object TSRMLS_CC));
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_regex_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Regex");
-			bson_append_regex(bson, key, key_len, php_phongo_regex_get_pattern(object TSRMLS_CC), php_phongo_regex_get_flags(object TSRMLS_CC));
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_javascript_ce TSRMLS_CC)) {
-			if (php_phongo_javascript_has_scope(object TSRMLS_CC)) {
-				mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Javascript with scope");
-				bson_append_code_with_scope(bson, key, key_len, php_phongo_javascript_get_code(object TSRMLS_CC), php_phongo_javascript_get_scope(object TSRMLS_CC));
-			} else {
-				mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Javascript without scope");
-				bson_append_code(bson, key, key_len, php_phongo_javascript_get_code(object TSRMLS_CC));
-			}
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_timestamp_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding Timestamp");
-			/*
-			 * WHOOPS! libbson swaps the increment/timestamp compared to BSON
-			 *
-			 * "Timestamp - Special internal type used by MongoDB replication and sharding.
-			 *  First 4 bytes are an increment, second 4 are a timestamp."
-			 */
-			bson_append_timestamp(bson, key, key_len, php_phongo_timestamp_get_timestamp(object TSRMLS_CC), php_phongo_timestamp_get_increment(object TSRMLS_CC));
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_maxkey_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding MaxKey");
-			bson_append_maxkey(bson, key, key_len);
-			return;
-		}
-		if (instanceof_function(Z_OBJCE_P(object), php_phongo_minkey_ce TSRMLS_CC)) {
-			mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding MinKey");
-			bson_append_minkey(bson, key, key_len);
-			return;
-		}
-
-		phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "Unexpected %s instance: %s", ZSTR_VAL(php_phongo_type_ce->name), ZSTR_VAL(Z_OBJCE_P(object)->name));
-		return;
-	} else {
-		HashTable *tmp_ht = HASH_OF(object);
-
-		if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-			ZEND_HASH_INC_APPLY_COUNT(tmp_ht);
-		}
-
-		mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "encoding document");
-		bson_append_document_begin(bson, key, key_len, &child);
-		phongo_zval_to_bson(object, flags, &child, NULL TSRMLS_CC);
-		bson_append_document_end(bson, &child);
-
-		if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-			ZEND_HASH_DEC_APPLY_COUNT(tmp_ht);
-		}
-	}
-}
-
-static void phongo_bson_append(bson_t *bson, php_phongo_bson_flags_t flags, const char *key, long key_len, zval *entry TSRMLS_DC)
-{
-#if PHP_VERSION_ID >= 70000
-try_again:
-#endif
-	switch (Z_TYPE_P(entry))
-	{
-		case IS_NULL:
-			bson_append_null(bson, key, key_len);
-			break;
-#if PHP_VERSION_ID >= 70000
-		case IS_TRUE:
-			bson_append_bool(bson, key, key_len, true);
-			break;
-
-		case IS_FALSE:
-			bson_append_bool(bson, key, key_len, false);
-			break;
-#else
-		case IS_BOOL:
-			bson_append_bool(bson, key, key_len, Z_BVAL_P(entry));
-			break;
-#endif
-
-		case IS_LONG:
-			BSON_APPEND_INT(bson, key, key_len, Z_LVAL_P(entry));
-			break;
-
-		case IS_DOUBLE:
-			bson_append_double(bson, key, key_len, Z_DVAL_P(entry));
-			break;
-
-		case IS_STRING:
-			if (bson_utf8_validate(Z_STRVAL_P(entry), Z_STRLEN_P(entry), true)) {
-				bson_append_utf8(bson, key, key_len, Z_STRVAL_P(entry), Z_STRLEN_P(entry));
-			} else {
-				phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "Got invalid UTF-8 value serializing '%s'", Z_STRVAL_P(entry));
-			}
-			break;
-
-		case IS_ARRAY:
-			if (php_phongo_is_array_or_document(entry TSRMLS_CC) == IS_ARRAY) {
-				bson_t     child;
-				HashTable *tmp_ht = HASH_OF(entry);
-
-				if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-					ZEND_HASH_INC_APPLY_COUNT(tmp_ht);
-				}
-
-				bson_append_array_begin(bson, key, key_len, &child);
-				phongo_zval_to_bson(entry, flags, &child, NULL TSRMLS_CC);
-				bson_append_array_end(bson, &child);
-
-				if (tmp_ht && ZEND_HASH_APPLY_PROTECTION(tmp_ht)) {
-					ZEND_HASH_DEC_APPLY_COUNT(tmp_ht);
-				}
-				break;
-			}
-			/* break intentionally omitted */
-		case IS_OBJECT:
-			object_to_bson(entry, flags, key, key_len, bson TSRMLS_CC);
-			break;
-
-#if PHP_VERSION_ID >= 70000
-		case IS_INDIRECT:
-			phongo_bson_append(bson, flags, key, key_len, Z_INDIRECT_P(entry) TSRMLS_DC);
-			break;
-
-		case IS_REFERENCE:
-			ZVAL_DEREF(entry);
-			goto try_again;
-#endif
-
-		default:
-			phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "Got unsupported type %d '%s'", Z_TYPE_P(entry), zend_get_type_by_const(Z_TYPE_P(entry)));
-	}
-}
-
-void phongo_zval_to_bson(zval *data, php_phongo_bson_flags_t flags, bson_t *bson, bson_t **bson_out TSRMLS_DC) /* {{{ */
-{
-	HashTable   *ht_data = NULL;
-#if PHP_VERSION_ID >= 70000
-	zval        obj_data;
-#else
-	HashPosition pos;
-	zval        *obj_data = NULL;
-#endif
-
-	/* If we will be encoding a class that may contain protected and private
-	 * properties, we'll need to filter them out later. */
-	bool         ht_data_from_properties = false;
-
-	switch(Z_TYPE_P(data)) {
-		case IS_OBJECT:
-			if (instanceof_function(Z_OBJCE_P(data), php_phongo_serializable_ce TSRMLS_CC)) {
-#if PHP_VERSION_ID >= 70000
-				zend_call_method_with_0_params(data, NULL, NULL, BSON_SERIALIZE_FUNC_NAME, &obj_data);
-#else
-				zend_call_method_with_0_params(&data, NULL, NULL, BSON_SERIALIZE_FUNC_NAME, &obj_data);
-#endif
-
-				if (Z_ISUNDEF(obj_data)) {
-					/* zend_call_method() failed */
-					break;
-				}
-
-#if PHP_VERSION_ID >= 70000
-				if (Z_TYPE(obj_data) != IS_ARRAY && !(Z_TYPE(obj_data) == IS_OBJECT && instanceof_function(Z_OBJCE(obj_data), zend_standard_class_def TSRMLS_CC))) {
-#else
-				if (Z_TYPE_P(obj_data) != IS_ARRAY && !(Z_TYPE_P(obj_data) == IS_OBJECT && instanceof_function(Z_OBJCE_P(obj_data), zend_standard_class_def TSRMLS_CC))) {
-#endif
-					phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC,
-						"Expected %s::%s() to return an array or stdClass, %s given",
-#if PHP_VERSION_ID >= 70000
-						Z_OBJCE_P(data)->name->val,
-#else
-						Z_OBJCE_P(data)->name,
-#endif
-						BSON_SERIALIZE_FUNC_NAME,
-#if PHP_VERSION_ID >= 70000
-						(Z_TYPE(obj_data) == IS_OBJECT
-							 ? Z_OBJCE(obj_data)->name->val
-							 : zend_get_type_by_const(Z_TYPE(obj_data))
-#else
-						(Z_TYPE_P(obj_data) == IS_OBJECT
-							 ? Z_OBJCE_P(obj_data)->name
-							 : zend_get_type_by_const(Z_TYPE_P(obj_data))
-#endif
-						)
-					);
-
-					break;
-				}
-
-#if PHP_VERSION_ID >= 70000
-				ht_data = HASH_OF(&obj_data);
-#else
-				ht_data = HASH_OF(obj_data);
-#endif
-
-				if (instanceof_function(Z_OBJCE_P(data), php_phongo_persistable_ce TSRMLS_CC)) {
-#if PHP_VERSION_ID >= 70000
-					bson_append_binary(bson, PHONGO_ODM_FIELD_NAME, -1, 0x80, (const uint8_t *)Z_OBJCE_P(data)->name->val, Z_OBJCE_P(data)->name->len);
-					zend_hash_str_del(ht_data, PHONGO_ODM_FIELD_NAME, sizeof(PHONGO_ODM_FIELD_NAME)-1);
-#else
-					bson_append_binary(bson, PHONGO_ODM_FIELD_NAME, -1, 0x80, (const uint8_t *)Z_OBJCE_P(data)->name, strlen(Z_OBJCE_P(data)->name));
-					zend_hash_del(ht_data, PHONGO_ODM_FIELD_NAME, sizeof(PHONGO_ODM_FIELD_NAME));
-#endif
-				}
-
-				break;
-			}
-
-			if (instanceof_function(Z_OBJCE_P(data), php_phongo_type_ce TSRMLS_CC)) {
-				phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "%s instance %s cannot be serialized as a root element", ZSTR_VAL(php_phongo_type_ce->name), ZSTR_VAL(Z_OBJCE_P(data)->name));
-
-				break;
-			}
-
-			ht_data = Z_OBJ_HT_P(data)->get_properties(data TSRMLS_CC);
-			ht_data_from_properties = true;
-			break;
-
-		case IS_ARRAY:
-			ht_data = HASH_OF(data);
-			break;
-
-		default:
-			return;
-	}
-
-	if (!ht_data || ZEND_HASH_GET_APPLY_COUNT(ht_data) > 1) {
-#if PHP_VERSION_ID >= 70000
-		if (Z_TYPE_P(data) == IS_OBJECT && instanceof_function(Z_OBJCE_P(data), php_phongo_serializable_ce TSRMLS_CC)) {
-#endif
-			if (!Z_ISUNDEF(obj_data)) {
-				zval_ptr_dtor(&obj_data);
-			}
-#if PHP_VERSION_ID >= 70000
-		}
-#endif
-		return;
-	}
-
-#if PHP_VERSION_ID >= 70000
-	{
-		zend_string *string_key = NULL;
-		zend_ulong   num_key = 0;
-		zval        *value;
-
-		ZEND_HASH_FOREACH_KEY_VAL(ht_data, num_key, string_key, value) {
-			if (string_key) {
-				if (ht_data_from_properties) {
-					/* Skip protected and private properties */
-					if (ZSTR_VAL(string_key)[0] == '\0' && ZSTR_LEN(string_key) > 0) {
-						zend_string_release(string_key);
-						continue;
-					}
-				}
-
-				if (strlen(ZSTR_VAL(string_key)) != ZSTR_LEN(string_key)) {
-					phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "BSON keys cannot contain null bytes. Unexpected null byte after \"%s\".", ZSTR_VAL(string_key));
-
-					return;
-				}
-
-				if (flags & PHONGO_BSON_ADD_ID) {
-					if (!strcmp(ZSTR_VAL(string_key), "_id")) {
-						flags &= ~PHONGO_BSON_ADD_ID;
-					}
-				}
-			}
-
-			/* Ensure we're working with a string key */
-			if (!string_key) {
-				string_key = zend_long_to_str(num_key);
-			} else {
-				zend_string_addref(string_key);
-			}
-
-			phongo_bson_append(bson, flags & ~PHONGO_BSON_ADD_ID, ZSTR_VAL(string_key), strlen(ZSTR_VAL(string_key)), value TSRMLS_CC);
-
-			zend_string_release(string_key);
-		} ZEND_HASH_FOREACH_END();
-	}
-#else
-	zend_hash_internal_pointer_reset_ex(ht_data, &pos);
-	for (;; zend_hash_move_forward_ex(ht_data, &pos)) {
-		char   *string_key = NULL;
-		uint    string_key_len = 0;
-		ulong   num_key = 0;
-		zval  **value;
-		int     hash_type;
-
-		hash_type = zend_hash_get_current_key_ex(ht_data, &string_key, &string_key_len, &num_key, 0, &pos);
-
-		if (hash_type == HASH_KEY_NON_EXISTENT) {
-			break;
-		}
-
-		if (zend_hash_get_current_data_ex(ht_data, (void **) &value, &pos) == FAILURE) {
-			break;
-		}
-
-		if (hash_type == HASH_KEY_IS_STRING) {
-			if (ht_data_from_properties) {
-				/* Skip protected and private properties */
-				if (string_key[0] == '\0' && string_key_len > 1) {
-					continue;
-				}
-			}
-
-			if (strlen(string_key) != string_key_len - 1) {
-				phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "BSON keys cannot contain null bytes. Unexpected null byte after \"%s\".", ZSTR_VAL(string_key));
-
-				return;
-			}
-
-			if (flags & PHONGO_BSON_ADD_ID) {
-				if (!strcmp(string_key, "_id")) {
-					flags &= ~PHONGO_BSON_ADD_ID;
-				}
-			}
-		}
-
-		/* Ensure we're working with a string key */
-		if (hash_type == HASH_KEY_IS_LONG) {
-			spprintf(&string_key, 0, "%ld", num_key);
-		}
-
-		phongo_bson_append(bson, flags & ~PHONGO_BSON_ADD_ID, string_key, strlen(string_key), *value TSRMLS_CC);
-
-		if (hash_type == HASH_KEY_IS_LONG) {
-			efree(string_key);
-		}
-	}
-#endif
-
-	if (flags & PHONGO_BSON_ADD_ID) {
-		bson_oid_t oid;
-
-		bson_oid_init(&oid, NULL);
-		bson_append_oid(bson, "_id", strlen("_id"), &oid);
-		mongoc_log(MONGOC_LOG_LEVEL_TRACE, MONGOC_LOG_DOMAIN, "Added new _id");
-		if (flags & PHONGO_BSON_RETURN_ID) {
-			if (bson_out) {
-				*bson_out = bson_new();
-				bson_append_oid(*bson_out, "_id", strlen("_id"), &oid);
-			}
-		}
-	}
-#if PHP_VERSION_ID >= 70000
-	if (Z_TYPE_P(data) == IS_OBJECT && instanceof_function(Z_OBJCE_P(data), php_phongo_serializable_ce TSRMLS_CC)) {
-#endif
-		if (!Z_ISUNDEF(obj_data)) {
-			zval_ptr_dtor(&obj_data);
-		}
-#if PHP_VERSION_ID >= 70000
-	}
-#endif
-}
-
-/* }}} */
-#if PHP_VERSION_ID >= 70000
-int phongo_bson_to_zval(const unsigned char *data, int data_len, zval *zv)
-#else
-int phongo_bson_to_zval(const unsigned char *data, int data_len, zval **zv)
+int php_phongo_bson_to_zval(const unsigned char *data, int data_len, zval **zv)
 #endif
 {
 	int retval = 0;
 	php_phongo_bson_state state = PHONGO_BSON_STATE_INITIALIZER;
 
-	retval = phongo_bson_to_zval_ex(data, data_len, &state);
+	retval = php_phongo_bson_to_zval_ex(data, data_len, &state);
 #if PHP_VERSION_ID >= 70000
 	ZVAL_ZVAL(zv, &state.zchild, 1, 1);
 #else
@@ -1504,9 +786,11 @@ int phongo_bson_to_zval(const unsigned char *data, int data_len, zval **zv)
 #endif
 
 	return retval;
-}
+} /* }}} */
 
-int phongo_bson_to_zval_ex(const unsigned char *data, int data_len, php_phongo_bson_state *state)
+/* Converts a BSON document to a PHP value according to the typemap specified in
+ * the state argument. The result will be set on the state argument. */
+int php_phongo_bson_to_zval_ex(const unsigned char *data, int data_len, php_phongo_bson_state *state) /* {{{ */
 {
 	      bson_reader_t *reader;
 	      bson_iter_t    iter;
@@ -1609,9 +893,9 @@ int phongo_bson_to_zval_ex(const unsigned char *data, int data_len, php_phongo_b
 	bson_reader_destroy(reader);
 
 	return 1;
-}
+} /* }}} */
 
-/* {{{ proto string BSON\fromPHP(array|object $value)
+/* {{{ proto string MongoDB\BSON\fromPHP(array|object $value)
    Returns the BSON representation of a PHP value */
 PHP_FUNCTION(fromPHP)
 {
@@ -1625,14 +909,13 @@ PHP_FUNCTION(fromPHP)
 	}
 
 	bson = bson_new();
-	phongo_zval_to_bson(data, PHONGO_BSON_NONE, bson, NULL TSRMLS_CC);
+	php_phongo_zval_to_bson(data, PHONGO_BSON_NONE, bson, NULL TSRMLS_CC);
 
 	PHONGO_RETVAL_STRINGL((const char *) bson_get_data(bson), bson->len);
 	bson_destroy(bson);
-}
-/* }}} */
+} /* }}} */
 
-static void apply_classname_to_state(const char *classname, int classname_len, php_phongo_bson_typemap_types *type, zend_class_entry **type_ce TSRMLS_DC)
+static void php_phongo_apply_classname_to_state(const char *classname, int classname_len, php_phongo_bson_typemap_types *type, zend_class_entry **type_ce TSRMLS_DC) /* {{{ */
 {
 	if (!strcasecmp(classname, "array")) {
 		*type = PHONGO_TYPEMAP_NATIVE_ARRAY;
@@ -1660,9 +943,11 @@ static void apply_classname_to_state(const char *classname, int classname_len, p
 			*type_ce = found_ce;
 		}
 	}
-}
+} /* }}} */
 
-void phongo_bson_typemap_to_state(zval *typemap, php_phongo_bson_typemap *map TSRMLS_DC)
+/* Applies the array argument to a typemap struct. An exception may be thrown if
+ * an invalid class name is provided. */
+void php_phongo_bson_typemap_to_state(zval *typemap, php_phongo_bson_typemap *map TSRMLS_DC) /* {{{ */
 {
 	if (typemap) {
 		char      *classname;
@@ -1671,7 +956,7 @@ void phongo_bson_typemap_to_state(zval *typemap, php_phongo_bson_typemap *map TS
 
 		classname = php_array_fetchc_string(typemap, "array", &classname_len, &classname_free);
 		if (classname_len) {
-			apply_classname_to_state(classname, classname_len, &map->array_type, &map->array TSRMLS_CC);
+			php_phongo_apply_classname_to_state(classname, classname_len, &map->array_type, &map->array TSRMLS_CC);
 		}
 		if (classname_free) {
 			str_efree(classname);
@@ -1679,7 +964,7 @@ void phongo_bson_typemap_to_state(zval *typemap, php_phongo_bson_typemap *map TS
 
 		classname = php_array_fetchc_string(typemap, "document", &classname_len, &classname_free);
 		if (classname_len) {
-			apply_classname_to_state(classname, classname_len, &map->document_type, &map->document TSRMLS_CC);
+			php_phongo_apply_classname_to_state(classname, classname_len, &map->document_type, &map->document TSRMLS_CC);
 		}
 		if (classname_free) {
 			str_efree(classname);
@@ -1687,15 +972,15 @@ void phongo_bson_typemap_to_state(zval *typemap, php_phongo_bson_typemap *map TS
 
 		classname = php_array_fetchc_string(typemap, "root", &classname_len, &classname_free);
 		if (classname_len) {
-			apply_classname_to_state(classname, classname_len, &map->root_type, &map->root TSRMLS_CC);
+			php_phongo_apply_classname_to_state(classname, classname_len, &map->root_type, &map->root TSRMLS_CC);
 		}
 		if (classname_free) {
 			str_efree(classname);
 		}
 	}
-}
+} /* }}} */
 
-/* {{{ proto array|object BSON\toPHP(string $bson [, array $typemap = array()])
+/* {{{ proto array|object MongoDB\BSON\toPHP(string $bson [, array $typemap = array()])
    Returns the PHP representation of a BSON value, optionally converting it into a custom class */
 PHP_FUNCTION(toPHP)
 {
@@ -1710,9 +995,9 @@ PHP_FUNCTION(toPHP)
 		return;
 	}
 
-	phongo_bson_typemap_to_state(typemap, &state.map TSRMLS_CC);
+	php_phongo_bson_typemap_to_state(typemap, &state.map TSRMLS_CC);
 
-	if (!phongo_bson_to_zval_ex((const unsigned char *)data, data_len, &state)) {
+	if (!php_phongo_bson_to_zval_ex((const unsigned char *)data, data_len, &state)) {
 		zval_ptr_dtor(&state.zchild);
 		RETURN_NULL();
 	}
@@ -1721,10 +1006,9 @@ PHP_FUNCTION(toPHP)
 #else
 	RETURN_ZVAL(state.zchild, 0, 1);
 #endif
-}
-/* }}} */
+} /* }}} */
 
-/* {{{ proto string BSON\toJSON(string $bson)
+/* {{{ proto string MongoDB\BSON\toJSON(string $bson)
    Returns the JSON representation of a BSON value */
 PHP_FUNCTION(toJSON)
 {
@@ -1767,10 +1051,9 @@ PHP_FUNCTION(toJSON)
 	}
 
 	bson_reader_destroy(reader);
-}
-/* }}} */
+} /* }}} */
 
-/* {{{ proto string BSON\fromJSON(string $json)
+/* {{{ proto string MongoDB\BSON\fromJSON(string $json)
    Returns the BSON representation of a JSON value */
 PHP_FUNCTION(fromJSON)
 {
@@ -1791,8 +1074,7 @@ PHP_FUNCTION(fromJSON)
 	} else {
 		phongo_throw_exception(PHONGO_ERROR_UNEXPECTED_VALUE TSRMLS_CC, "%s", error.domain == BSON_ERROR_JSON ? error.message : "Error parsing JSON");
 	}
-}
-/* }}} */
+} /* }}} */
 
 /*
  * Local variables:
